script/requirements.txt
google-genai>=1.0.0
feedparser>=6.0.11
